+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 0.2.0:
   * Make `Debug` string representations more compact.
   * `Month::wrapping_add/sub` now take an `i8` instead of `u8`.
   * Add `YearMonth::add/sub_years` and `YearMonth::add/sub_months`.
+  * Add `Date::add/sub_years` and `Date::add/sub_months`.
   * Add `InvalidDayOfMonth::next/prev_valid()` to get the nearest valid date.
 
 Version 0.1.1: